Add next button and fix PlayClip
I added a next button that will play the next song in the combo box. I don't know how this will work with shuffle but it might be better to shuffle the combobox contents instead of the playlist itself in the end.
Additionally, since the last patch the PlayClip function wasn't doing anything once you select another song. I provided a fix for the time being although I don't know if you plan on changing it based on how playlists work.